Examples folder cleaned

+++ SW/Click Examples mikroC/examples/8x8/eight_x_eight_hal.c
@@ -36,7 +36,7 @@
 *******************************************************************************/
 
 static void         ( *write_spi_p )            ( unsigned int data_out );
-static unsigned long( *read_spi_p )             ( unsigned int buffer );
+static unsigned ( *read_spi_p )             ( unsigned int buffer );
 
 extern sfr sbit LED_EIGHT_CS;
